requirements.txt
numpy==1.26.4
pandas
matplotlib
scikit-learn
seaborn
tensorflow
environs
apache-airflow==2.9.1
apache-airflow-providers-apache-kafka
gitpython
minio
mlflow
openpyxl